# Add benchmark allocation support and improve Emit table size