@agocke 2018-08-22
@dotnet/roslyn-compiler For a one line compiler change.