@agocke 2018-08-27
ping @dotnet/roslyn-compiler for review